Fix issues that was causing the Type-Writer to emit 'any' for lots of expression *not* of the 'any' type.
Specifically, while trying to find the containing decl, the typewriter looked for a non-null decl.
Unfortunately, the resolver returned 'undefined' in a lot of places, and hte typewriter accepted that
as non-null.

Fixing this caused a large number of typewriter baselines to change.  I did random spot checking of
all the cases where 'any' became a better type.  These all looked ok.  I also did checking of every
case where a non-'any' type became 'any'.  In all these cases, 'any' was the appropriate type, and
it was an improper decl that caused the non-any to be returned.