Update less tokens while parsing.
1) During normal parsing we don't need to do any updates at all.
2) During incremental parsing we only need to update the tokens after the edit.

This means there is no perf hit during normal parsing.

The average speed for incremental parsing (i.e. incremental parsing a large
file with edits in the middle of the file) was around 66MB/s before this change.
After this change, the average speed dropped to 55 MB/s.  That means a drop in
speed of only about 15%.

After this change there is no more need for AST.  Syntax nodes contain the
same positions as ast nodes.  And they also now contain parent pointers to
make working with them easier.

The only drawback is that the trees are no longer immutable.  Once incremental
parsing completes, the old tree is no longer usable.