Improve parser error recovery.
Jake ran into a repro from the F12 team where we ended up producing a pretty ugly parse tree
when we encountered certain constructs in an unterminated block.  Specifically, if something
like "interface Foo" was encountered we wouldn't properly bail out and parse that as an
interface.  This was due to the parsing askign the question "is this an expression" and
thinking "yes" due to 'interface' being a javascript identifier rather than a keyword.  We
need to be a bit less lenient here so that we can produce a more sensible tree for the rest
of the system to use.

Jake will be addressing hte crash that happens later in the system because of code that
wasn't expecting such a bizarre looking tree.