Bring parser expression parsing in line with the ES5 spec.
Previously, the expression parser contained a single method for parsing binary, conditional and
assignment expressions.  While this simplified things, it led to certain bugs going noticed in
expression parsing.  For example, the following was previously allowed by the parser:

    var a = () => {} || b;

This is not allowed because an arrow function is actually at the same precedence as an assignment
expression.  And, the left side of a logical or expression doesn't allow that precedence.

This change splits out the parsing code to handle the lowest precedence comma expressions, then
then assignment expresssion, then normal binary expressions.  Assignment expressions are called
out specially because (while they look like binary expressions) the grammar rules for them are
particularly different.

Lots of comments have also been added at many points to make it clear how parsing is supposed to
work, and how it maps to the ES5 grammar and the TS extensions.