Fixed missing case for noImplicitAny on function signatures.
One class of errors when using --noImplicitAny would not be reported
because it was reported during resolution, which would cause problems.

It appears that resolution caused some function signatures to be
considered typechecked before an implicit any could be detected/reported.
Here I have made moved some error-reporting code to the typechecking
phase, giving a (hopefully) cleaner separation of concerns.